Reorganize; remove cruft
