+++ spherical/grid/algebra.py
