+++ spherical/modes/algebra.py
