+++ spherical/modes/derivatives.py
