+++ spherical/modes/evaluate.py
