+++ spherical/modes/rotate.py
