+++ spherical/recursions/wignerH.py
@@ -46,41 +46,6 @@ def ϵ(m):
 
 
 @jit
-def wedge_size(ℓₘₐₓ, mpₘₐₓ=None):
-    """Total size of wedges of width mpₘₐₓ up to ℓₘₐₓ
-
-    Parameters
-    ----------
-    ℓₘₐₓ : int
-    mpₘₐₓ : int, optional
-        If None, it is assumed to be at least ℓ
-
-    See Also
-    --------
-    wedge_index : Index inside these wedges
-
-    Notes
-    -----
-    Here, it is assumed that only data with m≥|m'| are stored, and only
-    corresponding values are passed.  We also assume |m|≤ℓ and |m'|≤ℓ.  Neither of
-    these are checked.  The wedge array that this function indexes is ordered as
-
-        [
-            (ℓ, mp, m) for ℓ in range(ℓₘₐₓ+1)
-            for mp in range(-min(ℓ, mpₘₐₓ), min(ℓ, mpₘₐₓ)+1)
-            for m in range(abs(mp), ℓ+1)
-        ]
-
-    """
-    if ℓₘₐₓ < 0:
-        return 0
-    if mpₘₐₓ is None or mpₘₐₓ > ℓₘₐₓ:
-        return (ℓₘₐₓ+1) * (ℓₘₐₓ+2) * (2*ℓₘₐₓ+3) // 6
-    else:
-        return ((ℓₘₐₓ+1)*(ℓₘₐₓ+2)*(2*ℓₘₐₓ+3) - 2*(ℓₘₐₓ-mpₘₐₓ)*(ℓₘₐₓ-mpₘₐₓ+1)*(ℓₘₐₓ-mpₘₐₓ+2)) // 6
-
-
-@jit
 def _wedge_index(ℓ, mp, m, mpₘₐₓ):
     """Helper function for `wedge_index`"""
     i = wedge_size(ℓ-1, mpₘₐₓ)  # total size of everything with smaller ℓ
@@ -429,296 +394,3 @@ def _step_5(d, n_max, mp_max, Hwedge, Hv):
                       d6 * Hwedge[i+i2, j]
                     + d[i+i7] * Hwedge[i+i3, j]
                 )
-
-
-class HCalculator(object):
-    def __init__(self, n_max, mp_max=None):
-        """Object to repeatedly calculate Wigner H values
-
-        In particular the simplicity of the first two of these relations implies that
-        we only need to compute one fourth of the total number of elements.  There is
-        also a very accurate and efficient recursion method to compute these values.
-
-        Create this object using the largest value of `n` (also commonly denoted `j` or
-        `ℓ`) you expect to need, optionally create the `workspace` for a given shape of
-        cosβ using that method, and then call this object for a given value or array of
-        cosβ values.
-
-        The returned object is a series of "wedges" of the matrix, for the various
-        values of `n`, comprising just a quarter of the elements of the full matrix;
-        all remaining elements are determined by the first two symmetries above.  This
-        wedge has an initial dimension representing a multi-index for (ℓ, mp, m)
-        values, while following dimensions are just the same as cosβ.  Any required
-        value may be obtained with `wedge_index(ℓ, mp, m, mp_max)`.  The inner call
-        translates a general index into the equivalent index lying inside the wedge,
-        while the outer call translates that corrected (ℓ, mp, m) tuple into a linear
-        index into the array.
-
-        Example Usage
-        -------------
-        hcalc = HCalculator(n_max)
-        workspace = hcalc.workspace(cosβ)  # Note that cosβ can be an array of many values
-        # Possibly loop over many values of cosβ here
-        wedge = hcalc(cosβ, sinβ, workspace)
-
-        """
-        self.n_max = int(n_max)
-        if self.n_max < 0:
-            raise ValueError('Nonsensical value for n_max = {0}'.format(self.n_max))
-        if mp_max is not None:
-            self.mp_max = abs(int(mp_max))
-        else:
-            self.mp_max = self.n_max
-        if self.mp_max < 1:
-            raise ValueError(f'Currently require mp_max to be at least 1, not {self.mp_max}')
-        self.wedge_size = wedge_size(self.n_max, self.mp_max)
-        n = np.array([n for n in range(self.n_max+2) for m in range(-n, n+1)])
-        m = np.array([m for n in range(self.n_max+2) for m in range(-n, n+1)])
-        absn = np.array([n for n in range(self.n_max+2) for m in range(n+1)])
-        absm = np.array([m for n in range(self.n_max+2) for m in range(n+1)])
-        self.a = np.sqrt((absn+1+absm) * (absn+1-absm) / ((2*absn+1)*(2*absn+3)))
-        self.b = np.sqrt((n-m-1) * (n-m) / ((2*n-1)*(2*n+1)))
-        self.b[m<0] *= -1
-        self.d = 0.5 * np.sqrt((n-m) * (n+m+1))
-        self.d[m<0] *= -1
-        with np.errstate(divide='ignore', invalid='ignore'):
-            self.g = 2*(m+1) / np.sqrt((n-m)*(n+m+1))
-            self.h = np.sqrt((n+m+2)*(n-m-1) / ((n-m)*(n+m+1)))
-        if not (
-            np.all(np.isfinite(self.a)) and
-            np.all(np.isfinite(self.b)) and
-            np.all(np.isfinite(self.d))
-        ):
-            raise ValueError("Found a non-finite value inside this object")
-
-    def workspace(self, cosβ=[1.0]):
-        """Return a new workspace sized for cosβ.
-
-        Note that the particular values of cosβ do not matter at all; only the shape of
-        the array is used.  The returned array may be used repeatedly when calling this
-        object, and will be used as a work space.  This is obviously not thread-safe.
-
-        """
-        cosβ = np.asarray(cosβ, dtype=float)
-        return np.zeros((self.wedge_size+(self.n_max+1)**2+self.n_max+2,) + cosβ.shape, dtype=float)
-
-    def __call__(self, cosβ, sinβ=None, workspace=None):
-        """Compute a quarter of the H matrix
-
-        Parameters
-        ----------
-        cosβ : array_like
-            Values of cos(β) on which to evaluate the d matrix.
-        sinβ : array_like, optional
-            Values of sin(β) corresponding to the above.  If not given, this will be
-            computed automatically.
-        workspace : array_like, optional
-            A working array like the one returned by HCalculator.workspace for the
-            input cosβ.  If not present, a workspace will be created automatically.
-
-        Returns
-        -------
-        Hwedge : array
-            This is a 1-dimensional array of floats; see below.
-
-        See Also
-        --------
-        wigner_d : Compute the full Wigner d matrix
-        wigner_D : Compute the full Wigner 𝔇 matrix
-        rotate : Avoid computing the full 𝔇 matrix and rotate modes directly
-        evaluate : Avoid computing the full 𝔇 matrix and evaluate modes directly
-
-        Notes
-        -----
-        H is related to Wigner's (small) d via
-
-            dₗⁿᵐ = ϵₙ ϵ₋ₘ Hₗⁿᵐ,
-
-        where
-
-                 ⎧ 1 for k≤0
-            ϵₖ = ⎨
-                 ⎩ (-1)ᵏ for k>0
-
-        H has various advantages over d, including the fact that it can be efficiently
-        and robustly valculated via recurrence relations, and the following symmetry
-        relations:
-
-            H^{m', m}_n(β) = H^{m, m'}_n(β)
-            H^{m', m}_n(β) = H^{-m', -m}_n(β)
-            H^{m', m}_n(β) = (-1)^{n+m+m'} H^{-m', m}_n(π - β)
-            H^{m', m}_n(β) = (-1)^{m+m'} H^{m', m}_n(-β)
-
-        Because of these symmetries, we only need to evaluate at most 1/4 of all the
-        elements.
-
-        """
-        cosβ = np.asarray(cosβ, dtype=float)
-        if np.max(cosβ) > 1.0 or np.min(cosβ) < -1.0:
-            raise ValueError('Nonsensical value for range of cosβ: [{0}, {1}]'.format(np.min(cosβ), np.max(cosβ)))
-        cosβshape = cosβ.shape
-        cosβ = cosβ.ravel(order='K')
-        if sinβ is None:
-            sinβ = np.sqrt(1 - cosβ**2)
-        else:
-            if sinβ.shape != cosβshape:
-                raise ValueError(
-                    f"Input cosβ and sinβ must be the same shape; their shapes are {cosβshape} and {sinβ.shape}."
-                )
-            sinβ = sinβ.ravel(order='K')
-        workspace = workspace if workspace is not None else self.workspace(cosβ)
-        Hwedge = workspace[:self.wedge_size]
-        Hv = workspace[self.wedge_size:self.wedge_size+(self.n_max+1)**2]
-        Hextra = workspace[self.wedge_size+(self.n_max+1)**2:self.wedge_size+(self.n_max+1)**2+self.n_max+2]
-        _step_1(Hwedge)
-        _step_2(self.g, self.h, self.n_max, self.mp_max, Hwedge, Hextra, Hv, cosβ, sinβ)
-        _step_3(self.a, self.b, self.n_max, self.mp_max, Hwedge, Hextra, cosβ, sinβ)
-        _step_4(self.d, self.n_max, self.mp_max, Hwedge, Hv)
-        _step_5(self.d, self.n_max, self.mp_max, Hwedge, Hv)
-        Hwedge.reshape((-1,)+cosβshape)
-        return Hwedge
-
-    def wigner_d(self, cosβ, sinβ=None, workspace=None):
-        """Compute Wigner's d matrix dˡₘₚ,ₘ(β)
-
-        Parameters
-        ----------
-        cosβ : array_like
-            Values of cos(β) on which to evaluate the d matrix.
-        sinβ : array_like, optional
-            Values of sin(β) corresponding to the above.  If not given, this will be
-            computed automatically.
-        workspace : array_like, optional
-            A working array like the one returned by HCalculator.workspace for the
-            input cosβ.  If not present, a workspace will be created automatically.
-
-        Returns
-        -------
-        d : array
-            This is a 1-dimensional array of floats; see below.
-
-        See Also
-        --------
-        __call__ : Compute a portion of the H matrix
-        wigner_D : Compute the full Wigner 𝔇 matrix
-        rotate : Avoid computing the full 𝔇 matrix and rotate modes directly
-        evaluate : Avoid computing the full 𝔇 matrix and evaluate modes directly
-
-        Notes
-        -----
-        This function is the preferred method of computing the d matrix for large ell
-        values.  In particular, above ell≈32 standard formulas become completely
-        unusable because of numerical instabilities and overflow.  This function uses
-        stable recursion methods instead, and should be usable beyond ell≈1000.
-
-        The result is returned in a 1-dimensional array ordered as
-
-            [
-                d(ell, mp, m, β)
-                for ell in range(ell_max+1)
-                for mp in range(-ell, ell+1)
-                for m in range(-ell, ell+1)
-            ]
-
-        """
-        cosβ = np.asarray(cosβ, dtype=float)
-        ell_min = 0
-        ell_max = self.n_max
-        Hwedge = self(cosβ, sinβ, workspace)
-        d = np.empty((LMpM_total_size(ell_min, ell_max),) + cosβ.shape)
-        _fill_wigner_d(ell_min, ell_max, self.mp_max, d, Hwedge)
-        return d
-
-    def wigner_D(self, R, workspace=None):
-        """Compute Wigner's 𝔇 matrix
-
-        Parameters
-        ----------
-        R : array_like
-            Array to be interpreted as a quaternionic array (thus its final dimension
-            must have size 4), representing the rotations on which the 𝔇 matrix will be
-            evaluated.
-        workspace : array_like, optional
-            A working array like the one returned by HCalculator.workspace.  If not
-            present, a workspace will be created automatically.
-
-        Returns
-        -------
-        d : array
-            This is a 1-dimensional array of floats; see below.
-
-        See Also
-        --------
-        __call__ : Compute a portion of the H matrix
-        wigner_D : Compute the full Wigner 𝔇 matrix
-        rotate : Avoid computing the full 𝔇 matrix and rotate modes directly
-        evaluate : Avoid computing the full 𝔇 matrix and evaluate modes directly
-
-        Notes
-        -----
-        This function is the preferred method of computing the d matrix for large ell
-        values.  In particular, above ell≈32 standard formulas become completely
-        unusable because of numerical instabilities and overflow.  This function uses
-        stable recursion methods instead, and should be usable beyond ell≈1000.
-
-        This function computes 𝔇ˡₘₚ,ₘ(R).  The result is returned in a 1-dimensional
-        array ordered as
-
-            [
-                𝔇(ell, mp, m, R)
-                for ell in range(ell_max+1)
-                for mp in range(-ell, ell+1)
-                for m in range(-ell, ell+1)
-            ]
-
-        """
-        R = quaternionic.array(R)
-        z = R.to_euler_phases
-        ell_min = 0
-        ell_max = self.n_max
-        Hwedge = self(z[1].real, z[1].imag, workspace)
-        𝔇 = np.empty((LMpM_total_size(ell_min, ell_max),) + R.shape[:-1], dtype=complex)
-        zₐpowers = complex_powers(z[0], ell_max)
-        zᵧpowers = complex_powers(z[2], ell_max)
-        _fill_wigner_D(ell_min, ell_max, self.mp_max, 𝔇, Hwedge[:, 0], zₐpowers, zᵧpowers)
-        return 𝔇
-
-
-@jit
-def _fill_wigner_d(ell_min, ell_max, mp_max, d, Hwedge):
-    """Helper function for HCalculator.wigner_d"""
-    for ell in range(ell_min, ell_max+1):
-        for mp in range(-ell, ell+1):
-            for m in range(-ell, ell+1):
-                i_d = LMpM_index(ell, mp, m, ell_min)
-                i_H = wedge_index(ell, mp, m, mp_max)
-                d[i_d] = ϵ(mp) * ϵ(-m) * Hwedge[i_H]
-
-
-@jit
-def _fill_wigner_D(ell_min, ell_max, mp_max, 𝔇, Hwedge, zₐpowers, zᵧpowers):
-    """Helper function for HCalculator.wigner_D"""
-    # 𝔇ˡₘₚ,ₘ(R) = dˡₘₚ,ₘ(R) exp[iϕₐ(m-mp)+iϕₛ(m+mp)] = dˡₘₚ,ₘ(R) exp[i(ϕₛ+ϕₐ)m+i(ϕₛ-ϕₐ)mp]
-    # exp[iϕₛ] = R̂ₛ = hat(R[0] + 1j * R[3]) = zp
-    # exp[iϕₐ] = R̂ₐ = hat(R[2] + 1j * R[1]) = zm.conjugate()
-    # exp[i(ϕₛ+ϕₐ)] = zp * zm.conjugate() = z[2] = zᵧ
-    # exp[i(ϕₛ-ϕₐ)] = zp * zm = z[0] = zₐ
-    for ell in range(ell_min, ell_max+1):
-        for mp in range(-ell, 0):
-            for m in range(-ell, 0):
-                i_D = LMpM_index(ell, mp, m, ell_min)
-                i_H = wedge_index(ell, mp, m, mp_max)
-                𝔇[i_D] = ϵ(mp) * ϵ(-m) * Hwedge[i_H] * zᵧpowers[-m].conjugate() * zₐpowers[-mp].conjugate()
-            for m in range(0, ell+1):
-                i_D = LMpM_index(ell, mp, m, ell_min)
-                i_H = wedge_index(ell, mp, m, mp_max)
-                𝔇[i_D] = ϵ(mp) * ϵ(-m) * Hwedge[i_H] * zᵧpowers[m] * zₐpowers[-mp].conjugate()
-        for mp in range(0, ell+1):
-            for m in range(-ell, 0):
-                i_D = LMpM_index(ell, mp, m, ell_min)
-                i_H = wedge_index(ell, mp, m, mp_max)
-                𝔇[i_D] = ϵ(mp) * ϵ(-m) * Hwedge[i_H] * zᵧpowers[-m].conjugate() * zₐpowers[mp]
-            for m in range(0, ell+1):
-                i_D = LMpM_index(ell, mp, m, ell_min)
-                i_H = wedge_index(ell, mp, m, mp_max)
-                𝔇[i_D] = ϵ(mp) * ϵ(-m) * Hwedge[i_H] * zᵧpowers[m] * zₐpowers[mp]
